+++ pattern.hpp
@@ -5,10 +5,10 @@
 
 namespace pattern
 {
-static constexpr int scales = 4;
-static constexpr int orientations = 6;
+static constexpr size_t scales = 4;
+static constexpr size_t orientations = 6;
 
-static constexpr int window = 16;
+static constexpr size_t window = 16;
 static constexpr double alpha = 1.1;
 static constexpr double lambda = 2;
 static constexpr double sigma_u = 0.25;
